feat: Add support for loading tools from configuration file
This commit adds the ability to load tools from the configuration file. The `load_tools` function is called with the list of tools retrieved from the config file, and the tools are passed to the `initialize_agent` function. This allows the agent to have access to additional tools for providing more capabilities in conversations.

BREAKING CHANGE: `SerAPIWrapper` import has been removed since it is no longer needed.